move spoof param test so you can reload the other tests
